+++ src/index.ts
@@ -3,6 +3,5 @@ export { default as ExceptionHandler } from './exceptionHandler';
 export { default as FixtureGenerator } from './fixtureGenerator';
 export * from './middlewares';
 export { default as ResponseBuilder } from './responseBuilder';
-export * from './services';
 export * from './types';
 export * as Validator from 'express-validator';
